requirements.txt
requests
beautifulsoup4
pandas
PyYAML
lxml
python-dateutil
tqdm
schedule